net: Add NET_RANDOM_ETHADDR_OUI option
This patch adds a Kconfig option and associated
changes to net_random_ethaddr to allow an example
mac to be set in Kconfig that will provide the OUI
that will be used when a random mac address is created.
The local assignment bit is still set, however this
allows unprovisioned devices to boot with a predictable
hardware address that can be used for things like
PXE boot etc.

I am using a full mac as the input so existing parsing and
validation functions can be used.

Signed-off-by: Jon Nettleton <jon@solid-run.com>