imx: imx8mq_hb: Add USB Host support
This patch is important as currently Linux is not initializing the
USB clocks correctly, so if we don't do it u-boot USB host mode
does not work.  This is a temporary fix until the Linux kernel is
patched to work properly.

Regardless we want USB working properly for booting and recovering
our board, although this needs to be converted to a DM driver model.

Signed-off-by: Jon Nettleton <jon@solid-run.com>