phy: phy-fsl-imx8mq-usb: Add driver to handle USB3 phy
This is an adaption from the Linux driver that is used so
DWC3 can initialize the phy for the USB subsystem.

Note: this does not work with CONFIG_CLK but I have added
the necessary code so once the clock management is moved
to the DM model it will just work.  Currently it relies on
usb_clk_init() being called from the board file to initialize
the phy's clock.

Signed-off-by: Jon Nettleton <jon@solid-run.com>