imx: imx8m: allow binaries to be stored in a non-tree location
Currently the image scripts require all the firmware and binary
blobs to be stored in the root source tree for the image scripts
to find them.  This patch checks the BINDIR variable and will
use that to look for the binaries if it is set.  Otherwise
the old behaviour of using $srctree is used.

Signed-off-by: Jon Nettleton <jon@solid-run.com>