config_distro_bootcmd: add generic find_uuid command
This adds a generic find_uuid command that can be used to help
automate booting for generic medium distros that don't use
an initramfs and don't populate extlinux.conf during image creation.

This is only a helper and does nothing if the uuid is already set
in the environment, and overrides nothing if the extlinux.conf
has another root command specified.

Signed-off-by: Jon Nettleton <jon@solid-run.com>